Sponsors: fix logo URL handling & move UI
Make SponsorDetailDto.LogoUrl init-only and update GetSponsorDetailQueryHandler to return a new record with LogoUrl populated (avoids mutating an immutable record). Add Produces<CollectionResponse<SponsorSummaryResponse>>(200) to the ListActiveSponsors endpoint for correct response metadata. Move NebaSocialIcon.razor into Sponsors and update SocialNetwork.cs namespace to Neba.Website.Server.Sponsors.
